fix: Allow dynamic properties on test struct
