+++ tests/unit/Core/Framework/Struct/Fixture/StateStruct.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopware\Tests\Unit\Core\Framework\Struct\Fixture;
+
+use Shopware\Core\Framework\Struct\StateAwareTrait;
+
+/**
+ * @internal
+ */
+class StateStruct
+{
+    use StateAwareTrait;
+}
