+++ tests/unit/Core/Framework/Struct/Fixture/TestCollection.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopware\Tests\Unit\Core\Framework\Struct\Fixture;
+
+use Shopware\Core\Framework\Struct\Collection;
+
+/**
+ * @internal
+ *
+ * @template TElement
+ *
+ * @extends Collection<TElement>
+ */
+class TestCollection extends Collection
+{
+}
